+++ src/main/webapp/brand_category/brand.jsp
@@ -59,7 +59,7 @@
                                     <td><img src="${item.image}" alt=""></td>
                                     <td>
                                         <a href="http://localhost:8080/brands?action=edit&id=${item.id}"><button class="btn btn-edit"><i class="fa-solid fa-pen-to-square"></i></button></a>
-                                        <a href="http://localhost:8080/brands?action=delete&id=${item.id}"><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
+                                        <a href="http://localhost:8080/brands?action=delete&id=${item.id}" onclick="return confirm('Bạn có chắc chắn muốn xóa thương hiệu này không?')";><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
                                     </td>
                                 </tr>
                             </c:forEach>
