+++ src/main/webapp/department/department.jsp
@@ -61,7 +61,7 @@
                                 <td>${item.salaryCoefficient}</td>
                                 <td>
                                     <a href="http://localhost:8080/departments?action=edit&id=${item.id}"><button class="btn btn-edit"><i class="fa-solid fa-pen-to-square"></i></button></a>
-                                    <a href="http://localhost:8080/departments?action=delete&id=${item.id}"><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
+                                    <a href="http://localhost:8080/departments?action=delete&id=${item.id}" onclick="return confirm('Bạn có chắc chắn muốn xóa phòng ban này không?')";><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
                                 </td>
                             </tr>
                         </c:forEach>
