+++ src/main/webapp/product/product.jsp
@@ -80,7 +80,7 @@
                         <td class="large">${item.description}</td>
                         <td class="medium">
                             <a href="http://localhost:8080/products?action=edit&id=${item.id}"><button class="btn btn-edit"><i class="fa-solid fa-pen-to-square"></i></button></a>
-                            <a href="http://localhost:8080/products?action=delete&id=${item.id}"><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
+                            <a href="http://localhost:8080/products?action=delete&id=${item.id}" onclick="return confirm('Bạn có chắc chắn muốn xóa sản phẩm này không?')";><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
                         </td>
                     </tr>
                 </c:forEach>
