+++ src/main/webapp/role/role.jsp
@@ -59,7 +59,7 @@
                                 <td>${item.name}</td>
                                 <td>
                                     <a href="http://localhost:8080/role?action=edit&id=${item.id}"><button class="btn btn-edit"><i class="fa-solid fa-pen-to-square"></i></button></a>
-                                    <a href="http://localhost:8080/role?action=delete&id=${item.id}"><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
+                                    <a href="http://localhost:8080/role?action=delete&id=${item.id}" onclick="return confirm('Bạn có chắc chắn muốn xóa chức năng này không?')";><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
                                 </td>
                             </tr>
                         </c:forEach>
