+++ src/main/webapp/staff/staff.jsp
@@ -82,7 +82,7 @@
                         <td class="medium">${item.nameDepartment}</td>
                         <td class="medium">
                             <a href="http://localhost:8080/staff?action=edit&id=${item.id}"><button class="btn btn-edit"><i class="fa-solid fa-pen-to-square"></i></button></a>
-                            <a href="http://localhost:8080/staff?action=delete&id=${item.id}"><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
+                            <a href="http://localhost:8080/staff?action=delete&id=${item.id}" onclick="return confirm('Bạn có chắc chắn muốn xóa nhân sự này không?')";><button class="btn btn-remove"><i class="fa-solid fa-trash-can"></i></button></a>
                         </td>
                     </tr>
                 </c:forEach>
